Built Installers with Github Actions
Converts the travis build that created installers and deployed
to github releases to instead be a Github Actions. This build
action is run when the master branch is updated.

Context/Motivation: We are out of build minutes with Travis and
this restores the ability to build prereleases after each merge
to master.

Notes & Changes:
- The build number now comes from github actions, it is effectively
  reset and will be a 'low' number
- SIGNIFICANT: Instead of creating a new release every time we merge,
  there will now be just a single prerelease that will overwrite
  itself after each merge.

For follow-up:
- Need to determine a new mechanism for promoting pre-release builds
  to be releases. Likely this will be tagging and building a release
  branch.

Set build tag and build release branches